Add files via upload
place device on the group track of a group of a group. When a track inside the group is selected, all other tracks will be muted. With the 'ArmFollow' button enabled, the selected track will also become the only armed track in the group, if there is already an armed track in that group. While recording all functionality is paused, so that selecting a track in the middle of recording will not ruin a recording.